requirements.txt
selenium
bs4
psycopg2
apache-airflow